Rename static/ to public/; add layout.tsx file in app/; global footer is no longer a gatsby slice
