+++ tsconfig.json
@@ -108,9 +108,13 @@
   "display": "Recommended",
   "include": [
     "next-env.d.ts",
+    //"app/layout.tsx",
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts"
   ],
-  "exclude": ["node_modules"],
+  "exclude": [
+    "node_modules",
+    "app/globals.css",
+  ],
 }
